+++ Sources/SourceKitLSP/SourceKitLSPServer.swift
@@ -699,7 +699,6 @@ public actor SourceKitLSPServer {
 
   /// Send the given notification to the editor.
   public func sendNotificationToClient(_ notification: some NotificationType) {
-    logger.log("Sending notification: \(notification.forLogging)")
     client.send(notification)
   }
